Fix modal dismissal failure on browser navigation
If we have an open modal on turbolinks:load, we need
to reinitialize it and setup some custom behavior so
that it can be properly dismissed. This is typically
an issue when navigating +back+ from a facet filter
option that was displayed in a modal. Something about
Turbolinks caching seems to be disrupting the modal
event handling when loading the page from cache. I'm
Not sure it's worth trying to dive into the old
Bootstrap/jQuery code to figure out why it's not
playing nicely with Turbolinks, but this gets the job
done.

Addresses #47