Attach Borealis viewer initialization to 'turbolinks:load' event
Until now, we've been initializing the Borealis viewer on the
'DOMContentLoaded' event. This works really well for initial page loads,
but when we start considering browser navigation, it begins to break
down. This is because the content is loaded from the Turbolinks cache,
and no 'DOMContentLoaded' event is emitted in this scenario. If we
attach the listener to 'turbolinks:load' instead, we can successfully
initialize the player both when initially loading a document into it, as
well as navigating back to a previously loaded document.

Addresses #47