Add "Transcript" as an advanced search field
Transcripts are stored as child documents of a compound, and until now
have not been indexed in Solr. This change provides some customizations
to the CDMBL classes that transform the CDM data for indexing. Namely,
support for aggregating select fields of child documents onto an
indexed, multivalued field on the parent. The only concrete example of
this we have currently is for transcripts, but the design allows for
adding more via `MDL::CompoundAggregatingFieldTransformer#aggregations`

While most, if not all, of the CDMBL classes support dependency
injection of their collaborators, inheritance was still necessary here
because we couldn't drill dependencies down through the ETLWorker entry
point. For the future, I'd maybe consider updating CDMBL to support
something like this:

```ruby
CDMBL.configure do |config|
  config.field_transformer_class = CompoundAggregatingFieldTransformer
  config.transformer_class = CompoundAggregatingTransformer
  # etc
end
```

and have the CDMBL configuration class provide defaults. CDMBL classes
could still support DI for ease of testing, etc., but their defaults
could be read from the configuration singleton rather than be hard-coded
in the constructor's method signature. For example:

```ruby
module CDMBL
  class RecordTransformer
    attr_reader :record, :field_mappings, :field_transformer
    def initialize(record: {},
                   field_mappings: [],
                   field_transformer: CDMBL.configuration.field_transformer)
      @record            = record
      @field_mappings    = field_mappings
      @field_transformer = field_transformer
    end

    # ...
  end
end
```

This would allow us to remove the `lib/mdl/etl_worker.rb` file added in
this commit and replace it with an initializer that configures CDMBL as
described above. But before starting down that path, I want to validate
that this produces the desired search results.