Show all the collections on the self-service indexing page
For reasons I have yet to discover, Blacklight was limiting the facet
query used to generate the list of collections available to index. The
facet was configured in the `IndexingController` to have a limit of -1,
which removes the limit completely when used in `AdvancedController`,
but trying to set or remove a limit on the facet directly wasn't working
for me. Setting the `default_more_limit` on the blacklight config seems
to do the trick.