Generate static paths directly to cached thumbnails
We can eliminate a lot of work for both the browser and the Rails app by
generating thumbnail URL paths that directly reference the static files
in /public/assets/thumbnails rather than dynamic ones that go through
ThumbnailsController.

Additionally, this change migrates the default thumbnail images for
audio, video, and PDF documents as well as the 404 trainwreck image
directly into /public so that we can stop depending on AWS
infrastructure that we don't control.