Use an extended read timeout when calling CONTENTdm's OAI endpoint
CONTENTdm seems to take a very long time to respond when we use the
`from` parameter for selective harvesting. This has resulted in read
timeouts when using `Net::HTTP.get_response`, which has a default
timeout setting of 60 seconds and no way to provide an alternative. With
this change, we provide a different HTTP client to `OaiRequest`, which
still uses Net::HTTP, but significantly increases the read_timeout.